Blocks: Try to make MediaUpload component extensible (#5211)
* Blocks: Make MediaUpload compononent extensible

* Blocks: Add the placeholder for the Media Upload component implementation
