+++ blocks/api/index.js
@@ -4,7 +4,11 @@ export {
 	getPossibleBlockTransformations,
 	switchToBlockType,
 } from './factory';
-export { default as parse, getBlockAttributes } from './parser';
+export {
+	default as parse,
+	getBlockAttributes,
+	parseWithAttributeSchema,
+} from './parser';
 export { default as rawHandler } from './raw-handling';
 export {
 	default as serialize,
