+++ blocks/library/gallery/test/index.js
@@ -4,8 +4,6 @@
 import { name, settings } from '../';
 import { blockEditRender } from 'blocks/test/helpers';
 
-jest.mock( 'blocks/media-upload', () => () => '*** Mock(Media upload button) ***' );
-
 describe( 'core/gallery', () => {
 	test( 'block edit matches snapshot', () => {
 		const wrapper = blockEditRender( name, settings );
