+++ edit-post/hooks/index.js
@@ -1,5 +1,6 @@
 /**
  * Internal dependencies
  */
+import './blocks';
 import './more-menu';
 import './validate-use-once';
